read XLSX into POJO

+++ JavaTool/src/main/java/org/haoyi/entity/Point.java
@@ -1,22 +1,22 @@
 package org.haoyi.entity;
 
 public class Point {
-    private Float lat;
-    private Float lng;
+    private Double lat;
+    private Double lng;
 
-    public Float getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(Float lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
 
-    public Float getLng() {
+    public Double getLng() {
         return lng;
     }
 
-    public void setLng(Float lng) {
+    public void setLng(Double lng) {
         this.lng = lng;
     }
 }
